moves node_v1.go to v1/node pkg
