+++ api/pkg/handler/v1/node/node_test.go
@@ -1,4 +1,4 @@
-package handler_test
+package node_test
 
 import (
 	"context"
